Move tests into flocker.control.
